Refactor legal and pseudo-legal move generation
